NoiseReductionBase in lib-builtin-effects
